Huge mass of changes that should really have been a bunch of commits:
Use of `enum_from_primitive!()`.
Fleshed out cycle counting; read methods now return a (cycle, read) tuple, allowing us to use the cycle count only if we want. The `add_cycles!()` macro helps to make their use to be a bit nicer. Introduced an explicit Cycle struct.
`Interconnect` is overhauled, adding stubs for accessing the VRAM, palette, OAM and GamePak.
Thumb and arm cores now both generated from Python in a build script.
Started using `overflowing_{add,sub}` for setting overflow and carry flags, which use intrinsics meaning setv and setc instructions can simply be emitted for setting the flags.
Made `breakpoints` and `watchpoints` public on `Arm7TDMI` because there seemed to be no reason to contrain the interface for accessing them.
`Interconnect` now has the notion of a code segment, where it tracks the region of memory in which the PC is currently in.